skills/sanatani-astrology/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanatani-astrology"
version = "1.0.0"
description = "Vedic astrology (Jyotish) capabilities for OpenClaw"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"

dependencies = [
  "pyswisseph>=2.8.0",
  "pyephem>=4.1.0",
  "numpy>=1.21.0",
  "pandas>=1.3.0",
  "geopy>=2.2.0",
  "geonamescache>=1.0.0",
  "timezonefinder>=6.0.0",
  "pytz>=2021.3",
  "jsonschema>=4.0.0",
]

[project.optional-dependencies]
knowledge = ["chromadb>=0.4.22"]

[tool.setuptools.packages.find]
where = ["."]
include = ["sanatani_astrology*"]
